% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/complex_matrices.R
\name{as_adj_signed}
\alias{as_adj_signed}
\title{Convert a signed graph to a signed adjacency matrix}
\usage{
as_adj_signed(g, sparse = FALSE)
}
\arguments{
\item{g}{igraph object. Must have a "sign" edge attribute.}

\item{sparse}{Logical scalar, whether to return the result as a sparse matrix. The Matrix package is required for sparse matrices.}
}
\value{
signed adjacency matrix
}
\description{
This function returns the adjacency matrix for a signed graph
}
\seealso{
\link{as_adj_complex}
}
