% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/complex_matrices.R
\name{as_complex_adj}
\alias{as_complex_adj}
\title{Convert a signed graph to a complex adjacency matrix}
\usage{
as_complex_adj(g, attr)
}
\arguments{
\item{g}{igraph object}

\item{attr}{edge attribute name that encodes positve ("P"), negative ("N") and ambivalent ("A") ties.}
}
\value{
complex adjacency matrix
}
\description{
This function returns the adjacency matrix for a signed graph that contains ambivalent ties
}
