% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/complex_matrices.R
\name{as_complex_edges}
\alias{as_complex_edges}
\title{Convert Signed Network to Complex}
\usage{
as_complex_edges(g, attr = "type")
}
\arguments{
\item{g}{igraph object. Must have a "sign" edge attribute.}

\item{attr}{new edge attribute name that encodes positve ("P"), negative ("N") and ambivalent ("A") ties.}
}
\value{
igraph object
}
\description{
Convert Signed Network to Complex
}
\examples{
g <- sample_islands_signed(2,10,1,10)
as_complex_edges(g)
}
\author{
David Schoch
}
