% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/complex_matrices.R
\name{as_incidence_signed}
\alias{as_incidence_signed}
\title{Convert a signed two-mode network to a signed matrix}
\usage{
as_incidence_signed(g, sparse = FALSE)
}
\arguments{
\item{g}{igraph object (bipartite). Must have a "sign" edge attribute.}

\item{sparse}{Logical scalar, whether to return the result as a sparse matrix. The Matrix package is required for sparse matrices.}
}
\value{
signed incidence matrix
}
\description{
This function returns the incidence matrix for a signed two-mode network.
}
