% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_avatar.R
\docType{data}
\name{avatar}
\alias{avatar}
\title{Signed networks from Avatar: The Last Airbender}
\format{
igraph object
}
\source{
scraped from Avatar Wiki (https://avatar.fandom.com/wiki/Category:Characters)
}
\usage{
avatar
}
\description{
Allies/Enemy relations from Avatar: The Last Airbender
}
\keyword{datasets}
