% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/signed_triangles.R
\name{count_signed_triangles}
\alias{count_signed_triangles}
\title{count signed triangles}
\usage{
count_signed_triangles(g)
}
\arguments{
\item{g}{igraph object with a sign edge attribute.}
}
\value{
counts for all 4 signed triangle types
}
\description{
Counts the number of all possible signed triangles (+++),(++-), (+--) and (---)
}
\examples{
library(igraph)
g <- make_full_graph(4)
E(g)$sign <- c(-1, 1, 1, -1, -1, 1)
count_signed_triangles(g)
}
\seealso{
\link{signed_triangles}
}
\author{
David Schoch
}
