% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_cow.R
\docType{data}
\name{cowList}
\alias{cowList}
\title{Signed networks from Correlates of War}
\format{List of igraph objects}
\source{
http://mrvar.fdv.uni-lj.si/pajek/SVG/CoW/default.htm
}
\usage{
cowList
}
\description{
51 signed networks of inter state relations
}
\references{
Doreian, P. and Mrvar, A. (2015). "Structural Balance and Signed International Relations". \emph{Journal of Social Structure}, 16(2)
}
\keyword{datasets}
