% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/centrality_indices.R
\name{degree_signed}
\alias{degree_signed}
\title{Signed Degree}
\usage{
degree_signed(
  g,
  mode = c("all", "in", "out"),
  type = c("pos", "neg", "ratio", "net")
)
}
\arguments{
\item{g}{igraph object with a sign edge attribute.}

\item{mode}{character string, “out” for out-degree, “in” for in-degree or “all” for undirected networks.}

\item{type}{character string, “pos” or “neg” for counting positive or negative neighbors only,
"ratio" for pos/(pos+neg), or "net" for pos-neg.}
}
\value{
centrality scores as numeric vector.
}
\description{
several options to calculate the signed degree of vertices
}
\author{
David Schoch
}
