% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/random_graphs.R
\name{graph_circular_signed}
\alias{graph_circular_signed}
\title{circular signed graph}
\usage{
graph_circular_signed(n, r = 1, pos = 0.1, neg = 0.1)
}
\arguments{
\item{n}{number of nodes}

\item{r}{radius}

\item{pos}{distance fraction between positive edges}

\item{neg}{distance fraction between negative edges}
}
\value{
igraph graph
}
\description{
circular graph with positive and negative edges.
}
\examples{
library(igraph)
graph_circular_signed(n = 50)
}
\author{
David Schoch
}
