% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{graph_from_adjacency_matrix_signed}
\alias{graph_from_adjacency_matrix_signed}
\title{Create signed graphs from adjacency matrices}
\usage{
graph_from_adjacency_matrix_signed(A, mode = "undirected", ...)
}
\arguments{
\item{A}{square adjacency matrix of a signed graph}

\item{mode}{Character scalar, specifies how to interpret the supplied matrix.  Possible values are: directed, undirected}

\item{...}{additional parameters for \code{from_adjacency()}}
}
\value{
a signed network as igraph object
}
\description{
Create signed graphs from adjacency matrices
}
\examples{
A <- matrix(c(0, 1, -1, 1, 0, 1, -1, 1, 0), 3, 3)
graph_from_adjacency_matrix_signed(A)
}
