% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{is_signed}
\alias{is_signed}
\title{Check if network is a signed network}
\usage{
is_signed(g)
}
\arguments{
\item{g}{igraph object}
}
\value{
logical scalar
}
\description{
Check if network is a signed network
}
\examples{
g <- sample_islands_signed(2, 5, 1, 5)
is_signed(g)
}
