% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/complex_matrices.R
\name{laplacian_matrix_complex}
\alias{laplacian_matrix_complex}
\title{Complex Graph Laplacian}
\usage{
laplacian_matrix_complex(g, attr, norm = FALSE)
}
\arguments{
\item{g}{igraph object.}

\item{attr}{edge attribute name that encodes positve ("P"), negative ("N") and ambivalent ("A") ties.}

\item{norm}{Whether to calculate the normalized Laplacian. See definitions below.}
}
\value{
a complex matrix
}
\description{
The Laplacian of a signed graph containing ambivalent ties.
}
\details{
See \link[igraph]{laplacian_matrix} of igraph for more details. In the complex case, D is a diagonal matrix containing the absolute values of row sums of the complex adjacency matrix.
}
\seealso{
\link{laplacian_matrix_signed}
}
\author{
David Schoch
}
