% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/signed_triangles.R
\name{signed_triangles}
\alias{signed_triangles}
\title{count signed triangles}
\usage{
signed_triangles(g)
}
\arguments{
\item{g}{signed network.}
}
\value{
counts for all 4 signed triangle types
}
\description{
Counts the number of all possible signed triangles (+++),(++-), (+--) and (---)
}
\examples{
library(igraph)
g <- graph.full(4)
E(g)$sign <- c(-1,1,1,-1,-1,1)
signed_triangles(g)
}
\author{
David Schoch
}
