% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/signed_triangles.R
\name{triad_census_signed}
\alias{triad_census_signed}
\title{signed triad census}
\usage{
triad_census_signed(g)
}
\arguments{
\item{g}{igraph object with a sign edge attribute.}
}
\value{
counts for all 138 signed directed triangle types
}
\description{
triad census for signed graphs
}
\examples{
library(igraph)
g <- make_full_graph(4, directed = TRUE)
E(g)$sign <- rep(c(-1, 1, 1, -1, -1, 1), 2)
triad_census_signed(g)
}
\author{
David Schoch
}
