% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_tribes.R
\docType{data}
\name{tribes}
\alias{tribes}
\title{Florentine family marriage network}
\format{An igraph graph containing positive/negative ties among New Guinean highland tribes.}
\usage{
tribes
}
\description{
The data was downloaded from http://vlado.fmf.uni-lj.si/pub/networks/data/ucinet/gama.dat
}
\references{
Read, K. E. (1954) Cultures of the central highlands, New Guinea. \emph{Southwestern Journal of Anthropology}, 1–43.
}
\keyword{datasets}
