% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/balance_scores.R
\name{balance_score}
\alias{balance_score}
\title{balancedness of signed network}
\usage{
balance_score(g, method = "triangles")
}
\arguments{
\item{g}{signed network.}

\item{method}{string indicating the method to be used. See details for options}
}
\value{
balancedness score
}
\description{
Implements several indices to assess the balancedness of a network.
}
\details{
The method parameter can be one of
\describe{
\item{triangles}{Fraction of balanced triangles. Maximal if all triangles are balanced.}
\item{eigen}{}
}
}
\examples{
library(igraph)
g <- graph.full(4)
E(g)$sign <- c(-1,1,1,-1,-1,1)

balance_score(g, method = "triangles")
balance_score(g, method = "eigen")
}
\author{
David Schoch
}
