% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/balance_scores.R
\name{balance_score}
\alias{balance_score}
\title{balancedness of signed network}
\usage{
balance_score(g, method = "triangles")
}
\arguments{
\item{g}{signed network.}

\item{method}{string indicating the method to be used. See details for options}
}
\value{
balancedness score
}
\description{
Implements several indices to assess the balancedness of a network.
}
\details{
The method parameter can be one of
\describe{
\item{triangles}{Fraction of balanced triangles. Maximal (=1) if all triangles are balanced.}
\item{walk}{\eqn{\sum exp(\lambda_i) / \sum exp(\mu_i)}} where \eqn{\lambda_i} are the eigenvalues of the
signed adjacency matrix and \eqn{\mu_i} of the unsigned adjacency matrix. Maximal (=1) if all walks are balanced.
}
}
\examples{
library(igraph)
g <- graph.full(4)
E(g)$sign <- c(-1,1,1,-1,-1,1)

balance_score(g, method = "triangles")
balance_score(g, method = "eigen")
}
\references{
Estrada, E. (2019). Rethinking structural balance in signed social networks. \emph{Discrete Applied Mathematics}.
}
\author{
David Schoch
}
