% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/complex_matrices.R
\name{complex_walks}
\alias{complex_walks}
\title{Count Walks in complex signed network}
\usage{
complex_walks(g, attr, k)
}
\arguments{
\item{g}{igraph object.}

\item{attr}{edge attribute that encodes positive ("P"), negative ("N") and ambivalent ("A") ties.}

\item{k}{integer. length of walks}
}
\value{
igraph object
}
\description{
Count Walks in complex signed network
}
\examples{
g <- sample_islands_signed(2,10,1,10)
g <- as_complex_edges(g,attr="type")
complex_walks(g,attr="type",k = 3)
}
\author{
David Schoch
}
