% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{ggsigned}
\alias{ggsigned}
\title{Plot a signed or complex network}
\usage{
ggsigned(g, type = "signed", attr = NULL, edge_cols = NULL, weights = FALSE)
}
\arguments{
\item{g}{igraph object. Must have a "sign" edge attribute or an attribute containing "P", "N", "A"}

\item{type}{character string. either "signed" or "complex"}

\item{attr}{character string. edge attribute that containing "P", "N", "A" if type="complex"}

\item{edge_cols}{colors used for negative and positive (and ambivalent) ties}

\item{weights}{logical. If TRUE, weights are computed based on sign. Defaults to FALSE}
}
\value{
ggplot2 object
}
\description{
Plot a signed or complex network
}
\details{
This is a very rudimentary visualization of a signed network. If you are fluent in 'ggraph', you can probably cook up something more sophisticated. The function is thus mostly meant to give a quick overview of the network.
}
\author{
David Schoch
}
