% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blockmodel.R
\name{signed_blockmodel}
\alias{signed_blockmodel}
\title{Blockmodelling for signed networks}
\usage{
signed_blockmodel(g, k, alpha = 0.5)
}
\arguments{
\item{g}{signed graph}

\item{k}{number of clusters}

\item{alpha}{see details}
}
\value{
numeric vector of block assignments and the associated criterion value
}
\description{
Finds blocks of nodes with intra-positive and inter-negative edges
}
\details{
The function minimizes P(C)=\eqn{\alpha}P+(1-\eqn{\alpha})N,
where P is the total number of negative ties within plus-sets and P be the total number of
positive ties between plus-sets. So far only the structural balance model is implemented.
}
\references{
Doreian, Patrick and Andrej Mrvar (2009). Partitioning signed social networks. \emph{Social Networks} 31(1) 1-11
}
\author{
David Schoch
}
