% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/signed_triangles.R
\name{signed_triangles}
\alias{signed_triangles}
\title{list signed triangles}
\usage{
signed_triangles(g)
}
\arguments{
\item{g}{igraph object with a sign edge attribute.}
}
\value{
matrix of vertex ids and the number of positive ties per triangle
}
\description{
lists all possible signed triangles
}
\examples{
library(igraph)
g <- make_full_graph(4)
E(g)$sign <- c(-1, 1, 1, -1, -1, 1)
signed_triangles(g)
}
\seealso{
\link{count_signed_triangles}
}
\author{
David Schoch
}
