% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_tribes.R
\docType{data}
\name{tribes}
\alias{tribes}
\title{Signed network of New Guinean highland tribes}
\format{
An igraph object
}
\source{
http://vlado.fmf.uni-lj.si/pub/networks/data/ucinet/gama.dat
}
\usage{
tribes
}
\description{
Signed social network of tribes of the Gahuku–Gama alliance structure of the Eastern Central Highlands of New Guinea, from Kenneth Read.
The network contains sixteen tribes connected by friendship ("rova") and enmity ("hina").
}
\references{
Read, K. E. (1954) Cultures of the central highlands, New Guinea. \emph{Southwestern Journal of Anthropology}, 1–43.
}
\keyword{datasets}
